chore:  Improve tests reports (#3)
This pull request includes changes to the CI workflow and updates to the
`package.json` file to enhance testing and reporting capabilities. The
most important changes include adding permissions to the CI workflow and
modifying the test script to include a new test reporter.

Changes to CI workflow:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR14-R15):
Added write permissions for contents to the `build-and-test` job.

Updates to `package.json`:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L11-R14):
Modified the `test` script to use additional test reporters for better
output and reporting.
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R38):
Added the `node-test-github-reporter` package as a new dependency.